[FIX] Resolve 44 Pylance type errors across codebase
Type system improvements:
- StateManager: Added kill_switch_active property for external access
- BaseBrokerAPI: Added abstract disconnect() method with implementations
- ScreenerResult: Added passou property alias for compatibility
- Database: Fixed Union[str, Path] parameter and nullable return types
- Orchestrator: Initialize beta_multiplier before conditional use
- Watchdog: Updated handler type to support async callables

Type casting fixes:
- correlation.py: Explicit float() for pandas scalar operations
- risk_math.py: Explicit float() for numpy calculations

Nullable handling:
- ai_gateway.py: Handle nullable message.content
- judge.py: Handle nullable timestamp in cache retrieval

Test fixes:
- test_phase1.py: Use real ScreenerResult objects instead of incompatible mocks

External library type ignores (appropriate - no stubs available):
- telegram_bot.py: python-telegram-bot builder pattern
- news_aggregator.py: gnews library

+++ src/analysis/risk_math.py
@@ -107,7 +107,7 @@ class RiskCalculator:
             covariance = aligned.iloc[:, 0].cov(aligned.iloc[:, 1])
             variance = aligned.iloc[:, 1].var()
 
-            return covariance / variance if variance > 0 else 1.0
+            return float(covariance / variance) if variance > 0 else 1.0
 
         except Exception:
             return 1.0
@@ -120,7 +120,7 @@ class RiskCalculator:
 
     def _calculate_var(self, returns: pd.Series, confidence: float) -> float:
         """Calcula Value at Risk."""
-        return abs(np.percentile(returns, (1 - confidence) * 100)) * 100
+        return float(abs(np.percentile(returns, (1 - confidence) * 100))) * 100
 
     def _calculate_cvar(self, returns: pd.Series, confidence: float) -> float:
         """Calcula Conditional VaR (Expected Shortfall)."""
